requirements.txt
streamlit
pandas
Pillow
scikit-learn